She works. She's beautiful. She needs to be enclosed in the activity class. Goodbye SignUpDB.java
