+++ app/src/main/java/ca/rededaniskal/SignupActivity.java
@@ -1,12 +1,24 @@
 package ca.rededaniskal;
 
 import android.content.Intent;
+import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+
+import static android.content.ContentValues.TAG;
 
 public class SignupActivity extends AppCompatActivity {
 
@@ -81,7 +93,13 @@ public class SignupActivity extends AppCompatActivity {
 
     public void finalPass(){
         if (businessLogic.isValid()){
-            if(businessLogic.createNewUser()){
+            SignUpDB db = new SignUpDB();
+            String email = emailText.getText().toString();
+            String password = passwordText.getText().toString();
+
+            db.createUser(email, password);
+
+            if(db.isSuccess()){
                 /* TODO: Pass to new intent */
                 Toast.makeText(getApplicationContext(), "Sign Up Complete", Toast.LENGTH_SHORT).show();
             }else{
@@ -93,5 +111,63 @@ public class SignupActivity extends AppCompatActivity {
 
     }
 
+
+    public class SignUpDB {
+
+        // To read or write from the database, a database reference is needed
+        private DatabaseReference mDatabase;
+        private FirebaseAuth mAuth;
+        private boolean success;
+        private FirebaseUser newUser;
+
+        public SignUpDB(){
+            mDatabase = FirebaseDatabase.getInstance().getReference();
+
+            // Initialize FirebaseAuth
+            mAuth = FirebaseAuth.getInstance();
+            success = false;
+        }
+
+
+        public boolean isSuccess() {
+            return success;
+        }
+
+        public void setSuccess(boolean success) {
+            this.success = success;
+        }
+
+        public void createUser(String email, String password){
+
+
+            mAuth.createUserWithEmailAndPassword(email, password)
+                    .addOnCompleteListener(SignupActivity.this, new OnCompleteListener<AuthResult>() {
+                        @Override
+                        public void onComplete(@NonNull Task<AuthResult> task) {
+                            if (task.isSuccessful()) {
+                                // Sign in success, update UI with the signed-in user's information
+                                Log.d(TAG, "createUserWithEmail:success");
+                                FirebaseUser user = mAuth.getCurrentUser();
+                                //updateUI(user);
+                                success = true;
+                            } else {
+                                // If sign in fails, display a message to the user.
+                                Log.w(TAG, "createUserWithEmail:failure", task.getException());
+//                            updateUI(null);
+                            }
+
+                        }
+                    });
+        }
+
+        public FirebaseUser getNewUser() {
+            return newUser;
+        }
+
+        private void setNewUser(FirebaseUser newUser) {
+            this.newUser = newUser;
+        }
+    }
+    
 
 }
